fix : MainViewModel에서 memo들의 정보를 가지고 있도록 수정
- MainViewModel에서 필요한 memos를 관리하는 기능을 DetailModeViewModel에서 가져옴
- 사용하지 않게된 두개의 ViewModel 삭제
- 계속 사용하는 CalendarModeViewModel에서 MainViewModel과 겹치는 코드 삭제
- MainViewModel에서 observe하고 있는 변수가 변할 때, setCalendarMonthData를 부를 수 있도록 파라미터 수정
